+++ modules/config.py
@@ -1,0 +1,6 @@
+#config.py
+STACK = list() # [ ['{vari}', '{data}', {byte}], ...]
+VARIABLES = dict()
+idx_of_vari_tmp = 0
+commands = ["print", "push", "pop", "set", "clear", 'quit']
+CMD = ''
