--- 명령어 자동완성 test.py
@@ -1,31 +1,0 @@
-try:
-    import readline
-except ImportError:
-    import pyreadline as readline
-
-class MyCompleter(object):  # Custom completer
-
-    def __init__(self, options):
-        self.options = sorted(options)
-
-    def complete(self, text, state):
-        if state == 0:  # on first trigger, build possible matches
-            if text:  # cache matches (entries that start with entered text)
-                self.matches = [s for s in self.options 
-                                    if s and s.startswith(text)]
-            else:  # no text entered, all matches possible
-                self.matches = self.options[:]
-
-        # return match indexed by state
-        try: 
-            return self.matches[state]
-        except IndexError:
-            return None
-
-completer = MyCompleter(["print", "push", "pop", "set", "clear", 'quit'])
-readline.parse_and_bind('tab: complete')    
-readline.set_completer(completer.complete)
-
-while True:
-    input_ = input("Input: ")
-    print("You entered", input_)
